requirements.txt
# NAT Agentic Backend - light dependencies

# === Core API Server ===
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# === LLM Client ===
openai>=1.0.0
httpx>=0.26.0

# === Utilities ===
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0

# === Background Scheduling ===
apscheduler>=3.10.0

# === OpenAI Agents SDK ===
openai-agents>=0.9.0

# === Web Search ===
google-search-results>=2.4.2
ddgs>=9.0.0

# === Video / Frame Capture ===
opencv-python-headless>=4.9.0
